refactor(frontend): move App into app folder
