+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import App from "./App";
+import App from "./app/App";
 import { AuthProvider } from "./app/providers/AuthProvider";
 import AppQueryClientProvider from "./app/providers/QueryClientProvider";
 import "./index.css"; // Ensure global styles are applied
